+++ docs/docs.trychroma.com/markdoc/content/docs/overview/roadmap.md
@@ -10,11 +10,6 @@
 
 # Roadmap
 
-
-{% note type="default" title="Last Updated" %}
-`May 20, 2024`
-{% /note %}
-
 The goal of this doc is to align *core* and *community* efforts for the project and to share what's in store for this year!
 
 **Sections**
