+++ docs/docs.trychroma.com/markdoc/content/production/cloud-providers/aws.md
@@ -230,7 +230,7 @@ Add the `CHROMA_CLIENT_AUTH_CREDENTIALS` environment variable to your local envi
 export CHROMA_CLIENT_AUTH_CREDENTIALS="test-token"
 ```
 
-{% MarkdocTabs %}
+{% Tabs %}
 
 {% Tab label="python" %}
 
@@ -283,7 +283,7 @@ chromaClient.heartbeat()
 
 {% /Tab %}
 
-{% /MarkdocTabs %}
+{% /Tabs %}
 
 ## Observability with AWS
