+++ docs/docs.trychroma.com/markdoc/content/production/cloud-providers/gcp.md
@@ -216,7 +216,7 @@ Add the `CHROMA_CLIENT_AUTH_CREDENTIALS` environment variable to your local envi
 export CHROMA_CLIENT_AUTH_CREDENTIALS="test-token"
 ```
 
-{% MarkdocTabs %}
+{% Tabs %}
 
 {% Tab label="python" %}
 
@@ -269,7 +269,7 @@ chromaClient.heartbeat()
 
 {% /Tab %}
 
-{% /MarkdocTabs %}
+{% /Tabs %}
 
 ## Observability with GCP
